% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArcaneMissiles.R
\name{ArcaneMissiles}
\alias{ArcaneMissiles}
\title{Calculate the probability of Hearthstone in R.}
\usage{
ArcaneMissiles(enemyHP = c(3, 1, 1), damage = 1, times = 3,
  n = 10000)
}
\arguments{
\item{enemyHP}{Numeric, number of enemies on the field.Default is `c(3,1,1)`.}

\item{damage}{Numeric, damage caused by each one.Default is 1.}

\item{times}{Numeric, repeat times.Default is 3.}

\item{n}{Simulation times,default is 10000.}
}
\value{
If all inputs are integer and logical, then the output
  will be an probability.
}
\description{
`ArcaneMissiles()` returns the probability of of remaining enemies on
the field after using like Arcane Missiles in its arguments.
}
\details{
This is a generic function: A mechanism similar to arcane missiles
uses this function to calculate probabilities.
}
\examples{
#ArcaneMissiles(c(3,1,1),1,3)



\dontrun{
#ArcaneMissiles(enemyHP=c(3,1,1),times=6)
}


}
